fixed filter by user parking when logged out.
